Split AssertionMethodSelectionRuleTests into one file per analyzer
